Rename sanitized.rb to sanitized_filename.rb
